CS-10901: extract pure helpers out of prerenderer.ts
No behavior change. Three pure helpers extracted into their own files
so prerenderer.ts can stay focused on the Prerenderer orchestration
class. Each helper was already exported for unit tests; moving them
decouples the test harness from the Chrome-launching constructor.

  * async-semaphore.ts — the AsyncSemaphore primitive, used for the
    global render-concurrency cap (and, in CS-10946, per-affinity file
    admission control).
  * batch-ownership-gate.ts — the `computeBatchClearCacheGate` pure
    decision function + BatchOwner / BatchClearCacheDecision types
    for CS-10758 step 3.
  * render-settlement.ts — `decorateRenderErrorsWithTimings` free
    function + a new AffinitySnapshotSampler class that owns the
    peak-sampling state previously sprawled across five private
    fields/methods on Prerenderer (CS-10872's affinity-snapshot
    diagnostic).

prerenderer.ts loses ~285 lines (1071 → 786). The back-compat
`Prerenderer.decorateRenderErrorsWithTimings` static assignment
preserves the old access pattern so
`prerender-diagnostics-persistence-test.ts` and any external
consumers keep working unchanged.

Tests: `prerender-cancellation-test.ts` and
`prerender-batch-ownership-test.ts` now import directly from the new
files rather than from prerenderer.ts. 37/37 relevant tests pass.
